Broke more stuff, added a health bar and some code for it.

+++ FirstGame/Assets/CharacterController/PowerUp.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [CreateAssetMenu]
 public class PowerUp : ScriptableObject {
 
-	public int PowerLevel;
+	public float PowerLevel;
 
 	public MovePattern MovePattern;
